# backend/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Chat",
    "description": "API for the Team Challenge Chat project",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://teamchallenge-chat-jmsz.onrender.com"
    },
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/residential_complex": {
      "get": {
        "tags": [
          "Other"
        ],
        "summary": "Get short info about all complexes",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": "[]"
                }
              }
            },
            "description": "Success"
          }
        }
      },
      "post": {
        "description": "User must be an admin. Must have an ARRAY of entrances",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Other"
        ],
        "summary": "Create a new complex",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "images",
                  "parking",
                  "security",
                  "access_control",
                  "concierge",
                  "playground",
                  "closed_area",
                  "video_surveillance",
                  "description",
                  "floors",
                  "location",
                  "buildings"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "parking": {
                    "type": "boolean"
                  },
                  "security": {
                    "type": "boolean"
                  },
                  "access_control": {
                    "type": "boolean"
                  },
                  "concierge": {
                    "type": "boolean"
                  },
                  "playground": {
                    "type": "boolean"
                  },
                  "closed_area": {
                    "type": "boolean"
                  },
                  "video_surveillance": {
                    "type": "boolean"
                  },
                  "description": {
                    "type": "string"
                  },
                  "floors": {
                    "type": "number"
                  },
                  "location": {
                    "type": "string"
                  },
                  "buildings": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "address": {
                          "type": "string"
                        },
                        "entrances": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "number": {
                                "type": "number"
                              },
                              "apartment_min": {
                                "type": "number"
                              },
                              "apartment_max": {
                                "type": "number"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/residential_complex/{name}": {
      "get": {
        "tags": [
          "Other"
        ],
        "summary": "Get full info about a complex",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Name of the complex to get"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": "{\n\"_id\": \"66ed6047827428bc0be0bfef\",\n\"name\": \"Sunny Quarter\",\n\"images\": \"http://res.cloudinary.com/dtonpxhk7/image/upload/v1726832708/d63Bg4paJctlmRbv..png.png http://res.cloudinary.com/dtonpxhk7/image/upload/v1726832710/qfQo12Sh1frSjDWp..png.png \",\n\"parking\": true,\n\"security\": true,\n\"access_control\": true,\n\"concierge\": false,\n\"playground\": true,\n\"closed_area\": true,\n\"video_surveillance\": true,\n\"description\": \"Very cool complex\",\n\"floors\": 12,\n\"entrances\": 4,\n\"location\": \"Simferopolska, 2k, Dnipro, Dnipropetrovsk region \",\n\"sections\": 2,\n\"apartments\": 52\n}"
                }
              }
            },
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}